// VideoUploader/src/app/@common/enum.ts
export enum StatusCodes {
  Success = 200,
  Unauthorized = 401,
  InvalidToken = 402,
}

export enum ToastStatus {
  Primary = 'primary',
  Success = 'success',
  Warning = 'warning',
  Danger = 'danger',
  Info = 'info',
}

export enum DialogMode {
  Add = 1,
  Edit = 2,
  View = 3,
}

export enum UserType {
  Student = 1,
  Tutor = 2,
}
export enum ExamType {
  OLExamination = 1,
  ALExamination = 2,
}
export enum PopMenuType {
  Profile = 1,
  LogOut = 2,
}